+++ Makefile
@@ -6,12 +6,12 @@
 #    By: toshi <toshi@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/06/15 03:56:20 by toshi             #+#    #+#              #
-#    Updated: 2024/08/28 22:23:17 by toshi            ###   ########.fr        #
+#    Updated: 2024/08/29 04:08:16 by toshi            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 CC			:=	cc -fsanitize=address
-# CFLAGS		:=	-Wall -Wextra -Werror
+CFLAGS		:=	-Wall -Wextra -Werror
 RM			:=	rm -rf
 
 NAME		:=	cub3d
@@ -64,17 +64,18 @@ SRCS		:=	$(SRCS_DIR)/$(CALC_DIR)/calc_calc_horizontal_inter_pos.c\
 				$(SRCS_DIR)/initialize_mlx.c \
 				$(SRCS_DIR)/main.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_add_color.c \
+				$(SRCS_DIR)/$(PARSE_DIR)/parse_add_wall.c \
+				$(SRCS_DIR)/$(PARSE_DIR)/parse_double_strdup_padd_space.c \
+				$(SRCS_DIR)/$(PARSE_DIR)/parse_extract_val.c \
+				$(SRCS_DIR)/$(PARSE_DIR)/parse_get_line_kind.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_load_map_list_to_data.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_load_map_str_to_list.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_load_textures.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_trim_map_list.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_utils_bool.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_utils_free.c \
-				$(SRCS_DIR)/$(PARSE_DIR)/parse_utils_load.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_utils_print.c \
-				$(SRCS_DIR)/$(PARSE_DIR)/parse_double_strdup_padd_space.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse_validate_map_data.c \
-				$(SRCS_DIR)/$(PARSE_DIR)/parse_utils.c \
 				$(SRCS_DIR)/$(PARSE_DIR)/parse.c \
 				$(SRCS_DIR)/$(MLX_WRAPPER_DIR)/ft_x_mlx_init.c \
 				$(SRCS_DIR)/$(MLX_WRAPPER_DIR)/ft_x_mlx_new_image.c \
