+++ srcs/parse/parse_get_line_kind.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_utils_load.c                                 :+:      :+:    :+:   */
+/*   parse_get_line_kind.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: toshi <toshi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/13 14:36:36 by tyamauch          #+#    #+#             */
-/*   Updated: 2024/08/28 22:08:14 by toshi            ###   ########.fr       */
+/*   Updated: 2024/08/29 04:08:36 by toshi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ static bool	is_map_str(char *str)
 	while (*str && *str != '\n')
 	{
 		if (!is_map_element(*str)
-			|| player_flag && is_player(*str))
+			|| (player_flag && is_player(*str)))
 			return (false);
 		if (is_player(*str))
 			player_flag = true;
@@ -51,21 +51,4 @@ t_line_kind	get_line_kind(char *str)
 	if (ft_strncmp(str, "C ", 2) == 0)
 		return (KIND_CEILING);
 	return (KIND_FALSE);
-}
-
-char	*extract_val(char *str, t_line_kind kind)
-{
-	size_t	identifer_len;
-	char	*val_include_space;
-	char	*val;
-
-	if (is_wall(kind))
-		identifer_len = 2;
-	else
-		identifer_len = 1;
-	val_include_space = ft_x_substr(str, identifer_len,
-			ft_strlen(str) - identifer_len);
-	val = ft_x_strtrim(val_include_space, " \t\n");
-	free(val_include_space);
-	return (val);
 }
