+++ srcs/parse/parse_load_textures.c
@@ -6,30 +6,11 @@
 /*   By: toshi <toshi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/13 14:33:34 by tyamauch          #+#    #+#             */
-/*   Updated: 2024/08/26 14:40:07 by toshi            ###   ########.fr       */
+/*   Updated: 2024/08/29 04:01:12 by toshi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parse.h"
-
-void		add_color(t_loader *loader, t_line_kind kind, char *str);
-
-static void	add_wall(t_loader *loader, t_line_kind kind, char *str)
-{
-	char	*val;
-
-	val = extract_val(str, kind);
-	if (!ft_can_open(val, O_RDONLY))
-		ft_my_perror_exit("The texture path is incorrect", 1);
-	if (kind == KIND_NORTH)
-		loader->north_path = val;
-	else if (kind == KIND_SOUTH)
-		loader->south_path = val;
-	else if (kind == KIND_WEST)
-		loader->west_path = val;
-	else if (kind == KIND_EAST)
-		loader->east_path = val;
-}
 
 static bool	is_member_default(t_loader loader, t_line_kind kind)
 {
